Breaking Change - default to string headers. No discovery.
